requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==44.0.2
Django==4.2.16
django-allauth==65.7.0
django-ckeditor-5==0.2.17
django-crispy-forms==2.3
django-jalali-date==2.0.0
django-taggit==6.1.0
humanize==4.10.0
idna==3.10
jalali-core==1.0.0
jdatetime==5.0.0
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
